moved http metrics to http layer
