+++ app/http/middleware/MetricsMiddleware.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/kosatnkn/catalyst/app/metrics"
+	"github.com/kosatnkn/catalyst/app/http/metrics"
 )
 
 // MetricsMiddleware attaches metrics to the request.
