--- app/metrics/initializer.go
@@ -1,8 +1,0 @@
-package metrics
-
-import "github.com/prometheus/client_golang/prometheus"
-
-// Init registers declared metrics.
-func Init() {
-	prometheus.MustRegister(HTTPReqDuration)
-}
